Organized project into folders. Implemented DB
