--- DisplayInfo.ui
@@ -1,242 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>DisplayInfo</class>
- <widget class="QDialog" name="DisplayInfo">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>1449</width>
-    <height>772</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>Dialog</string>
-  </property>
-  <widget class="QPushButton" name="PrintStadiumPushButton">
-   <property name="geometry">
-    <rect>
-     <x>0</x>
-     <y>410</y>
-     <width>271</width>
-     <height>31</height>
-    </rect>
-   </property>
-   <property name="font">
-    <font>
-     <pointsize>14</pointsize>
-    </font>
-   </property>
-   <property name="text">
-    <string>Print All Stadiums with Teams</string>
-   </property>
-  </widget>
-  <widget class="QPushButton" name="PrintNFCTeamsPushButton">
-   <property name="geometry">
-    <rect>
-     <x>0</x>
-     <y>280</y>
-     <width>271</width>
-     <height>31</height>
-    </rect>
-   </property>
-   <property name="font">
-    <font>
-     <pointsize>14</pointsize>
-    </font>
-   </property>
-   <property name="text">
-    <string>Print NFC Teams</string>
-   </property>
-  </widget>
-  <widget class="QPushButton" name="PrintStadiumByDatePushButton">
-   <property name="geometry">
-    <rect>
-     <x>0</x>
-     <y>470</y>
-     <width>271</width>
-     <height>31</height>
-    </rect>
-   </property>
-   <property name="font">
-    <font>
-     <pointsize>14</pointsize>
-    </font>
-   </property>
-   <property name="text">
-    <string>Print Stadium By Date</string>
-   </property>
-  </widget>
-  <widget class="QPushButton" name="PrintAllTeamsPushButton">
-   <property name="geometry">
-    <rect>
-     <x>0</x>
-     <y>160</y>
-     <width>271</width>
-     <height>31</height>
-    </rect>
-   </property>
-   <property name="font">
-    <font>
-     <pointsize>14</pointsize>
-    </font>
-   </property>
-   <property name="text">
-    <string>Print All Teams</string>
-   </property>
-  </widget>
-  <widget class="QPushButton" name="PrintAFCTeamsPushButton">
-   <property name="geometry">
-    <rect>
-     <x>0</x>
-     <y>220</y>
-     <width>271</width>
-     <height>31</height>
-    </rect>
-   </property>
-   <property name="font">
-    <font>
-     <pointsize>14</pointsize>
-    </font>
-   </property>
-   <property name="text">
-    <string>Print AFC Teams</string>
-   </property>
-  </widget>
-  <widget class="QPushButton" name="PrintNorthNFCTeamsPushButton">
-   <property name="geometry">
-    <rect>
-     <x>0</x>
-     <y>340</y>
-     <width>271</width>
-     <height>31</height>
-    </rect>
-   </property>
-   <property name="font">
-    <font>
-     <pointsize>14</pointsize>
-    </font>
-   </property>
-   <property name="text">
-    <string>Print North NFC Teams</string>
-   </property>
-  </widget>
-  <widget class="QTableWidget" name="tableWidget">
-   <property name="geometry">
-    <rect>
-     <x>340</x>
-     <y>20</y>
-     <width>1101</width>
-     <height>741</height>
-    </rect>
-   </property>
-  </widget>
-  <widget class="QLabel" name="label">
-   <property name="geometry">
-    <rect>
-     <x>10</x>
-     <y>30</y>
-     <width>241</width>
-     <height>41</height>
-    </rect>
-   </property>
-   <property name="font">
-    <font>
-     <pointsize>14</pointsize>
-    </font>
-   </property>
-   <property name="text">
-    <string>Select the City You Want</string>
-   </property>
-  </widget>
-  <widget class="QPushButton" name="ShowBermudaGrassStadiumsPushButton">
-   <property name="geometry">
-    <rect>
-     <x>0</x>
-     <y>590</y>
-     <width>271</width>
-     <height>31</height>
-    </rect>
-   </property>
-   <property name="font">
-    <font>
-     <pointsize>14</pointsize>
-    </font>
-   </property>
-   <property name="text">
-    <string>Show Bermuda Grass Stadiums</string>
-   </property>
-  </widget>
-  <widget class="QPushButton" name="ShowOpenRoofStadiumsPushButton">
-   <property name="geometry">
-    <rect>
-     <x>0</x>
-     <y>530</y>
-     <width>271</width>
-     <height>31</height>
-    </rect>
-   </property>
-   <property name="font">
-    <font>
-     <pointsize>14</pointsize>
-    </font>
-   </property>
-   <property name="text">
-    <string>Show Open Roof Stadiums</string>
-   </property>
-  </widget>
-  <widget class="QPushButton" name="PrintTeamInforByConferencePushButton">
-   <property name="geometry">
-    <rect>
-     <x>0</x>
-     <y>650</y>
-     <width>271</width>
-     <height>31</height>
-    </rect>
-   </property>
-   <property name="font">
-    <font>
-     <pointsize>14</pointsize>
-    </font>
-   </property>
-   <property name="text">
-    <string>Print Team Info By Conference</string>
-   </property>
-  </widget>
-  <widget class="QPushButton" name="HomePushButton">
-   <property name="geometry">
-    <rect>
-     <x>0</x>
-     <y>710</y>
-     <width>271</width>
-     <height>51</height>
-    </rect>
-   </property>
-   <property name="font">
-    <font>
-     <pointsize>14</pointsize>
-    </font>
-   </property>
-   <property name="text">
-    <string>Home</string>
-   </property>
-  </widget>
-  <widget class="QComboBox" name="comboBox">
-   <property name="geometry">
-    <rect>
-     <x>0</x>
-     <y>80</y>
-     <width>311</width>
-     <height>41</height>
-    </rect>
-   </property>
-   <property name="font">
-    <font>
-     <pointsize>14</pointsize>
-    </font>
-   </property>
-  </widget>
- </widget>
- <resources/>
- <connections/>
-</ui>
