--- Header.h
@@ -1,17 +1,0 @@
-#ifndef HEADER_H
-#define HEADER_H
-
-//For QClasses
-#include <QMainWindow>
-#include <QDialog>
-#include <QMessageBox>
-#include <QDebug>
-
-//For Background Image
-#include <QPixmap>
-
-//For RNG
-#include <QRandomGenerator>
-#include <QTime>
-
-#endif // HEADER_H
