--- PrintByTeam.cpp.autosave
@@ -1,14 +1,0 @@
-#include "PrintByTeam.h"
-#include "ui_DisplayInfo.h"
-
-DisplayInfo::DisplayInfo(QWidget *parent) :
-    QDialog(parent),
-    ui(new Ui::DisplayInfo)
-{
-    ui->setupUi(this);
-}
-
-DisplayInfo::~DisplayInfo()
-{
-    delete ui;
-}
