+++ app/Header.h
@@ -1,0 +1,17 @@
+#ifndef HEADER_H
+#define HEADER_H
+
+//For QClasses
+#include <QMainWindow>
+#include <QDialog>
+#include <QMessageBox>
+#include <QDebug>
+
+//For Background Image
+#include <QPixmap>
+
+//For RNG
+#include <QRandomGenerator>
+#include <QTime>
+
+#endif // HEADER_H
