+++ db/souvenirs.csv
@@ -1,0 +1,160 @@
+1,1,Signed Helmets,77.99
+2,1,Autographed Football,99.89
+3,1,Team Pennant,17.99
+4,1,Team Picture,29.99
+5,1,Team Jersey,199.99
+6,2,Signed Helmets,77.99
+7,2,Autographed Football,99.89
+8,2,Team Pennant,17.99
+9,2,Team Picture,29.99
+10,2,Team Jersey,199.99
+11,3,Signed Helmets,77.99
+12,3,Autographed Football,99.89
+13,3,Team Pennant,17.99
+14,3,Team Picture,29.99
+15,3,Team Jersey,199.99
+16,4,Signed Helmets,77.99
+17,4,Autographed Football,99.89
+18,4,Team Pennant,17.99
+19,4,Team Picture,29.99
+20,4,Team Jersey,199.99
+21,5,Signed Helmets,77.99
+22,5,Autographed Football,99.89
+23,5,Team Pennant,17.99
+24,5,Team Picture,29.99
+25,5,Team Jersey,199.99
+26,6,Signed Helmets,77.99
+27,6,Autographed Football,99.89
+28,6,Team Pennant,17.99
+29,6,Team Picture,29.99
+30,6,Team Jersey,199.99
+31,7,Signed Helmets,77.99
+32,7,Autographed Football,99.89
+33,7,Team Pennant,17.99
+34,7,Team Picture,29.99
+35,7,Team Jersey,199.99
+36,8,Signed Helmets,77.99
+37,8,Autographed Football,99.89
+38,8,Team Pennant,17.99
+39,8,Team Picture,29.99
+40,8,Team Jersey,199.99
+41,9,Signed Helmets,77.99
+42,9,Autographed Football,99.89
+43,9,Team Pennant,17.99
+44,9,Team Picture,29.99
+45,9,Team Jersey,199.99
+46,10,Signed Helmets,77.99
+47,10,Autographed Football,99.89
+48,10,Team Pennant,17.99
+49,10,Team Picture,29.99
+50,10,Team Jersey,199.99
+51,11,Signed Helmets,77.99
+52,11,Autographed Football,99.89
+53,11,Team Pennant,17.99
+54,11,Team Picture,29.99
+55,11,Team Jersey,199.99
+56,12,Signed Helmets,77.99
+57,12,Autographed Football,99.89
+58,12,Team Pennant,17.99
+59,12,Team Picture,29.99
+60,12,Team Jersey,199.99
+61,13,Signed Helmets,77.99
+62,13,Autographed Football,99.89
+63,13,Team Pennant,17.99
+64,13,Team Picture,29.99
+65,13,Team Jersey,199.99
+66,14,Signed Helmets,77.99
+67,14,Autographed Football,99.89
+68,14,Team Pennant,17.99
+69,14,Team Picture,29.99
+70,14,Team Jersey,199.99
+71,15,Signed Helmets,77.99
+72,15,Autographed Football,99.89
+73,15,Team Pennant,17.99
+74,15,Team Picture,29.99
+75,15,Team Jersey,199.99
+76,16,Signed Helmets,77.99
+77,16,Autographed Football,99.89
+78,16,Team Pennant,17.99
+79,16,Team Picture,29.99
+80,16,Team Jersey,199.99
+81,17,Signed Helmets,77.99
+82,17,Autographed Football,99.89
+83,17,Team Pennant,17.99
+84,17,Team Picture,29.99
+85,17,Team Jersey,199.99
+86,18,Signed Helmets,77.99
+87,18,Autographed Football,99.89
+88,18,Team Pennant,17.99
+89,18,Team Picture,29.99
+90,18,Team Jersey,199.99
+91,19,Signed Helmets,77.99
+92,19,Autographed Football,99.89
+93,19,Team Pennant,17.99
+94,19,Team Picture,29.99
+95,19,Team Jersey,199.99
+96,20,Signed Helmets,77.99
+97,20,Autographed Football,99.89
+98,20,Team Pennant,17.99
+99,20,Team Picture,29.99
+100,20,Team Jersey,199.99
+101,21,Signed Helmets,77.99
+102,21,Autographed Football,99.89
+103,21,Team Pennant,17.99
+104,21,Team Picture,29.99
+105,21,Team Jersey,199.99
+106,22,Signed Helmets,77.99
+107,22,Autographed Football,99.89
+108,22,Team Pennant,17.99
+109,22,Team Picture,29.99
+110,22,Team Jersey,199.99
+111,23,Signed Helmets,77.99
+112,23,Autographed Football,99.89
+113,23,Team Pennant,17.99
+114,23,Team Picture,29.99
+115,23,Team Jersey,199.99
+116,24,Signed Helmets,77.99
+117,24,Autographed Football,99.89
+118,24,Team Pennant,17.99
+119,24,Team Picture,29.99
+120,24,Team Jersey,199.99
+121,25,Signed Helmets,77.99
+122,25,Autographed Football,99.89
+123,25,Team Pennant,17.99
+124,25,Team Picture,29.99
+125,25,Team Jersey,199.99
+126,26,Signed Helmets,77.99
+127,26,Autographed Football,99.89
+128,26,Team Pennant,17.99
+129,26,Team Picture,29.99
+130,26,Team Jersey,199.99
+131,27,Signed Helmets,77.99
+132,27,Autographed Football,99.89
+133,27,Team Pennant,17.99
+134,27,Team Picture,29.99
+135,27,Team Jersey,199.99
+136,28,Signed Helmets,77.99
+137,28,Autographed Football,99.89
+138,28,Team Pennant,17.99
+139,28,Team Picture,29.99
+140,28,Team Jersey,199.99
+141,29,Signed Helmets,77.99
+142,29,Autographed Football,99.89
+143,29,Team Pennant,17.99
+144,29,Team Picture,29.99
+145,29,Team Jersey,199.99
+146,30,Signed Helmets,77.99
+147,30,Autographed Football,99.89
+148,30,Team Pennant,17.99
+149,30,Team Picture,29.99
+150,30,Team Jersey,199.99
+151,31,Signed Helmets,77.99
+152,31,Autographed Football,99.89
+153,31,Team Pennant,17.99
+154,31,Team Picture,29.99
+155,31,Team Jersey,199.99
+156,32,Signed Helmets,77.99
+157,32,Autographed Football,99.89
+158,32,Team Pennant,17.99
+159,32,Team Picture,29.99
+160,32,Team Jersey,199.99
