+++ db/teamInfo.csv
@@ -1,0 +1,32 @@
+1,Arizona Cardinals,State Farm Stadium,"63,400","Glendale, Arizona",National Football Conference,NFC West,Bermuda Grass,Retractable,2026
+2,Atlanta Falcons,Mercedes-Benz Stadium,"71,000","Atlanta, Georgia",National Football Conference,NFC South,FieldTurf Classic HD,Fixed,2017
+3,Baltimore Ravens,M&T Bank Stadium,"71,008","Baltimore, Maryland",American Football Conference,AFC North,Bermuda Grass,Open,1998
+4,Buffalo Bills,Bills Stadiuim,"71,608","Orchard Park, New York",American Football Conference,AFC East,A-Turf Titan 50,Open,1973
+5,Carolina Panthers,Bank of America Stadium,"75,419","Charlotte, North Carolina",National Football Conference,NFC South,Bermuda Grass,Open,1996
+6,Chicago Bears,Soldier Field,"61,500","Chicago, Illinois",National Football Conference,NFC North,Kentucky Bluegrass,Open,1924
+7,Cincinnati Bengals,Paul Brown Stadium,"65,515","Cincinnati, Ohio",American Football Conference,AFC North,UBU Speed Series S5-M Synthetic Turf,Open,2000
+8,Cleveland Browns,FirstEnergy Stadium,"67,895","Cleveland, Ohio",American Football Conference,AFC North,Kentucky Bluegrass,Open,1999
+9,Dallas Cowboys,AT&T Stadium,"80,000","Arlington, Texas",National Football Conference,NFC East,Hellas Matrix Turf,Retractable,2009
+10,Denver Broncos,Empower Field at Mile High,"76,125","Denver, Colorado",American Football Conference,AFC West,Kentucky Bluegrass,Open,2001
+11,Detroit Lions,Ford Field,"65,000","Detroit, Michigan",National Football Conference,NFC North,FieldTurf Classic HD,Fixed,2002
+12,Green Bay Packers,Lambeau Field,"81,435","Green Bay, Wisconsin",National Football Conference,NFC North,Desso GrassMaster,Open,1957
+13,Houston Texans,NRG Stadium,"72,220","Houston, Texas",American Football Conference,AFC South,Hellas Matrix Turf,Retractable,2002
+14,Indianapolis Colts,Lucas Oil Stadium,"67,000","Indianapolis, Indiana",American Football Conference,AFC South,FieldTurf Classic HD,Retractable,2008
+15,Jacksonville Jaguars,TIAA Bank Field,"69,132","Jacksonville, Florida",American Football Conference,AFC South,Bermuda Grass,Open,1995
+16,Kansas City Chiefs,Arrowhead Stadium,"76,416","Kansas City, Missouri",American Football Conference,AFC West,Bermuda Grass,Open,1972
+17,Las Vegas Raiders,Allegiant Stadium ,"65,000","Paradise, Nevada",American Football Conference,AFC West,Bermuda Grass,Fixed,2020
+18,Los Angeles Chargers,SoFi Stadium,"70,240","Inglewood, California",American Football Conference,AFC West,Bermuda Grass,Fixed,2020
+19,Los Angeles Rams,SoFi Stadium,"70,240","Inglewood, California",National Football Conference,NFC West,Bermuda Grass,Fixed ,2020
+20,Miami Dolphins,Hard Rock Stadium,"65,326","Miami Gardens, Florida",American Football Conference,AFC East,Platinum TE Paspalum,Open,1987
+21,Minnesota Vikings,U.S. Bank Stadium,"66,655","Minneapolis, Minnesota",National Football Conference,NFC North,UBU Speed Series S5-M Synthetic Turf,Fixed,2016
+22,New England Patriots,Gillette Stadium,"66,829","Foxborough, Massachusetts",American Football Conference,AFC East,FieldTurfÿRevolution,Open,2002
+23,New Orleans Saints,Mercedes-Benz Superdome,"73,000","New Orleans, Louisiana",National Football Conference,NFC South,UBU Sports Speed Series S5-M Synthetic Turf,Fixed,1975
+24,New York Giants,MetLife Stadium,"82,500","East Rutherford, New Jersey",National Football Conference,NFC East,UBU Sports Speed Series S5-M Synthetic Turf,Open,2010
+25,New York Jets,MetLife Stadium,"82,500","East Rutherford, New Jersey",American Football Conference,AFC East,UBU Sports Speed Series S5-M Synthetic Turf,Open,2010
+26,Philadelphia Eagles,Lincoln Financial Field,"69,596","Philadelphia, Pennsylvania",National Football Conference,NFC East,Desso GrassMaster,Open,2003
+27,Pittsburgh Steelers,Heinz Field,"68,400","Pittsburgh, Pennsylvania",American Football Conference,AFC North,Kentucky Bluegrass,Open,2001
+28,San Francisco 49ers,Levi's Stadium,"68,500","Santa Clara, California",National Football Conference,NFC West,Bermuda grass  ,Open,2014
+29,Seattle Seahawks,CenturyLink Field,"68,000","Seattle, Washington",National Football Conference,NFC West,FieldTurfÿRevolution,Open,2002
+30,Tampa Bay Buccaneers,Raymond James Stadium,"65,890","Tampa, Florida",National Football Conference,NFC South,Bermuda Grass,Open,1998
+31,Tennessee Titans,Nissan Stadium,"69,143","Nashville, Tennessee",American Football Conference,AFC South,Bermuda Grass,Open,1999
+32,Washington Redskins,FedExField,"82,000","Landover, Maryland",National Football Conference,NFC East,Bermuda Grass,Open,1997
